+++ like/views.py
@@ -8,7 +8,7 @@ from .models import Like
 
 
 # 改变点赞状态
-@method_decorator(login_required, name='dispatch')
+# @method_decorator(login_required, name='dispatch')
 class LikeChangeView(TemplateView):
     template_name = None
     context = {}
@@ -17,14 +17,15 @@ class LikeChangeView(TemplateView):
         model = self.request.GET.get('content_type')
         content_type = ContentType.objects.get(model=model)
         object_id = self.request.GET.get('object_id')
-        liker = self.request.user
-
-        like, created = Like.objects.get_or_create(content_type=content_type, object_id=object_id, liker=liker)
-        if created:
-            self.context['status'] = 'CREATED_SUCCESS'
+        if self.request.user.is_authenticated:
+            like, created = Like.objects.get_or_create(content_type=content_type, object_id=object_id, liker=self.request.user)
+            if created:
+                self.context['status'] = 'CREATED_SUCCESS'
+            else:
+                like.delete()
+                self.context['status'] = 'DELETE_SUCCESS'
         else:
-            like.delete()
-            self.context['status'] = 'DELETE_SUCCESS'
+            self.context['status'] = 'NO_LOGIN_ERROR'
         return self.context
 
     def get(self, request, *args, **kwargs):
